requirements.txt
# langchain
# langchain-core
# langchain-community
# langchain-groq

python-dotenv
ipykernel

pypdf
bs4
pymupdf==1.22.3
wikipedia
langchain-text-splitters

chromadb
sentence-transformers
faiss-cpu

pandas
mysql-connector-python
SQLAlchemy
validators==0.28.1
unstructured>=0.12.0
numexpr

streamlit
fastapi
numba>=0.58.0
llvmlite>=0.41.0

# LangChain ecosystem (compatible versions)
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-text-splitters==0.2.4
